turn into javascript for module compatibility
