+++ src/posts/index.js
@@ -26,7 +26,7 @@ require('./bookmarks')(Posts);
 require('./queue')(Posts);
 require('./diffs')(Posts);
 require('./uploads')(Posts);
-require('./endorse.ts')(Posts);
+require('./endorse')(Posts);
 
 Posts.exists = async function (pids) {
     return await db.exists(
